requirements.txt
pysha3==1.0.2
pycoin==0.80